% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_LaTeX2.R
\name{as_LaTeX2}
\alias{as_LaTeX2}
\alias{latex2}
\title{Coerce to LaTeX2}
\usage{
as_LaTeX2(x)

latex2(...)
}
\arguments{
\item{x}{An object to convert to a \link{LaTeX2} object.}

\item{...}{Objects to concatenate.}
}
\value{
\code{as_LaTeX2()} converts \code{x} to a \link{LaTeX2} object.

\code{latex2()} converts the arguments to \link{LaTeX2} objects
and concatenates them into a new \link{LaTeX2} object.
}
\description{
Coerce to LaTeX2
}
