% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{brace_options}
\alias{brace_options}
\title{Extract brace wrapped options from Latex code.}
\usage{
brace_options(items, which = 1, drop = TRUE)
}
\arguments{
\item{items}{A list of latex items}

\item{which}{Return this block}

\item{drop}{Should whitespace be dropped from the result?}
}
\description{
Some Latex environments and macros take optional parameters
wrapped in curly brackets (braces).  This function extracts those
if they immediately follow the environment or macro and possibly
some bracketed options.
}
