% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bracket_options}
\alias{bracket_options}
\title{Extract bracket wrapped options from Latex code.}
\usage{
bracket_options(items, which = 1, drop = TRUE, start = 1)
}
\arguments{
\item{items}{A list of latex items}

\item{which}{Which bracket options do you want?  Some
macros support more than one set.}

\item{drop}{Should whitespace be dropped?}

\item{start}{Start looking at `items[[start]]`}
}
\description{
Many Latex environments and macros take optional parameters
wrapped in square brackets.  This function extracts those,
assuming they come immediately after the macro.
}
