% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLatex.R
\name{deparseLatex}
\alias{deparseLatex}
\title{Convert latex object into character vector}
\usage{
deparseLatex(x, dropBraces = FALSE)
}
\arguments{
\item{x}{A latex object}

\item{dropBraces}{Whether to drop unnecessary braces.}
}
\value{
A character vector.
}
\description{
Convert latex object into character vector
}
