% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_brace_options}
\alias{find_brace_options}
\title{Find brace wrapped options in Latex code.}
\usage{
find_brace_options(items, which = 1, start = 1)
}
\arguments{
\item{items}{A list of latex items}

\item{which}{Return this block}

\item{start}{Start looking at `items[[start]]`}
}
\value{
The index of the block containing the options.
}
\description{
Some Latex environments and macros take optional parameters
wrapped in curly brackets (braces).  This function finds those
if they immediately follow the environment or macro and possibly
some bracketed options.
}
