% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_bracket_options}
\alias{find_bracket_options}
\title{Find bracket wrapped options from Latex code.}
\usage{
find_bracket_options(items, which = 1, start = 1)
}
\arguments{
\item{items}{A list of latex items}

\item{which}{Which bracket options do you want?  Some
macros support more than one set.}

\item{start}{Start looking at `items[[start]]`}
}
\value{
Indices into `items` of the options (including the
brackets)
}
\description{
Many Latex environments and macros take optional parameters
wrapped in square brackets.  This function finds those,
assuming they come immediately after the macro.
}
