% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_contents}
\alias{get_contents}
\alias{set_contents}
\title{Convenience functions to get or set contents of item}
\usage{
get_contents(item)

set_contents(item, value)
}
\arguments{
\item{item}{An item from a Latex list (or a \link{LaTeX2} list with one item).}

\item{value}{An object that can be coerced to be
a \link{LaTeX2} object.}
}
\value{
\code{get_contents} returns the contents of the item as a \link{LaTeX2} list.

\code{set_contents} returns the original \code{item} with the contents
replaced by \code{value}.
}
\description{
Convenience functions to get or set contents of item
}
\examples{
get_contents(parseLatex("{abc}"))

set_contents(parseLatex("{abc}"), "def")
}
