% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{find_bracket_options}
\alias{find_bracket_options}
\alias{find_brace_options}
\title{Find macro or environment options}
\usage{
find_bracket_options(items, which = 1, start = 1)

find_brace_options(items, which = 1, start = 1)
}
\arguments{
\item{items}{A list of latex items}

\item{which}{Return this block}

\item{start}{Start looking at `items[[start]]`}
}
\value{
`find_bracket_options` returns indices into `items` of the options (including the
brackets)

`find_brace_options` returns the index of the block containing the options.
}
\description{
Many Latex environments and macros take optional parameters
wrapped in square brackets.  `find_bracket_options` finds those,
assuming they come immediately after the macro.

Some Latex environments and macros take optional parameters
wrapped in curly brackets (braces). `find_brace_options` finds those
if they immediately follow the environment or macro (and possibly
some bracketed options).
}
