% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLatex.R
\docType{package}
\name{parseLatex_pkg}
\alias{parseLatex-package}
\alias{parseLatex_pkg}
\title{The parseLatex package}
\description{
Exports an enhanced version of the tools::parseLatex() function to handle 'LaTeX' syntax more accurately. Also includes numerous functions for searching and modifying 'LaTeX' source.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmurdoch/parseLatex}
  \item \url{https://dmurdoch.github.io/parseLatex/}
  \item Report bugs at \url{https://github.com/dmurdoch/parseLatex/issues}
}

}
\author{
\strong{Maintainer}: Duncan Murdoch \email{murdoch.duncan@gmail.com}

Authors:
\itemize{
  \item Duncan Murdoch \email{murdoch.duncan@gmail.com}
}

Other contributors:
\itemize{
  \item The R Core Team [contributor, copyright holder]
}

}
