% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{path_to}
\alias{path_to}
\title{Find path to a particular kind of item}
\usage{
path_to(items, all = FALSE, is_fn, ...)
}
\arguments{
\item{items}{A list of latex items.}

\item{all}{Find all matching, or the first?}

\item{is_fn}{Which test function to use.}

\item{...}{Additional parameters to pass to `is_fn`.}
}
\value{
`path_to()` returns the recursive path to the
first example matching the `is_fn` conditions,
or a list of paths to all matching items.
}
\description{
Find path to a particular kind of item
}
\details{
`path_to()` does a recursive search in the order
items appear in the deparse.
}
\examples{
latex <- kableExtra::kbl(mtcars[1:2, 1:2], format = "latex", caption = "Sample table")
parsed <- parseLatex(latex)
parsed
path <- path_to(parsed, is_fn = is_env,
                        envtypes = "tabular")
parsed[[path]]
}
