% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLatex.R
\name{print.LaTeX2item}
\alias{print.LaTeX2item}
\alias{print.LaTeX2}
\title{Print methods}
\usage{
\method{print}{LaTeX2item}(x, ...)

\method{print}{LaTeX2}(x, tags = FALSE, ...)
}
\arguments{
\item{x}{Object to work on.}

\item{...}{Extra parameters to pass to \link{deparseLatex}.}

\item{tags}{Whether to display LaTeX2 tags.}
}
\description{
Print methods
}
