% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcrefs.R
\name{rmSrcrefs}
\alias{rmSrcrefs}
\title{Remove srcrefs}
\usage{
rmSrcrefs(items)
}
\arguments{
\item{items}{A \link{LaTeX2} object, or any other list of \link{LaTeX2item}s.}
}
\value{
The \code{items} with source references removed.
}
\description{
Remove srcrefs
}
