% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableRow.R
\name{split_table}
\alias{split_table}
\alias{split_row}
\title{Split up a table by rows or columns}
\usage{
split_table(table)

split_row(row)
}
\arguments{
\item{table}{A tabular-like environment to work with.}

\item{row}{A list of items from a single row of a table.}
}
\value{
A \link{LaTeX2item} object which is the same table
with an \link{ITEMLIST} holding the rows.  The attribute
\code{has_itemlist} will be set to \code{TRUE}.

A \link{LaTeX2item} object which is the same row
with an \link{ITEMLIST} holding the cells.  The attribute
\code{has_itemlist} will be set to \code{TRUE}.
}
\description{
Split up a table by rows or columns
}
\examples{
latex <- kableExtra::kbl(mtcars[1:2, 1:2], format = "latex")
parsed <- parseLatex(latex)
table <- split_table(parsed[[find_tabular(parsed)]])
print(latex2(table), tags = TRUE)
row <- split_row(tableRow(table, 2))
print(latex2(row), tags = TRUE)
}
