% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitting}
\alias{splitting}
\alias{split_list}
\alias{split_latex}
\title{Splitting lists of items}
\usage{
split_list(items, splits, include = FALSE)

split_latex(...)
}
\arguments{
\item{items}{A \link{LaTeX2} or similar list.}

\item{splits}{Which item numbers divide the parts?}

\item{include}{If \code{TRUE}, include the split item at
the end of each part.}

\item{...}{Arguments to pass to \code{split_list}.}
}
\value{
\code{split_list()} returns a list of pieces
separated at the splits.

\code{split_latex()} returns a list of pieces
separated at the splits.  Each piece is marked as
an \code{ITEMLIST} item, and the whole thing is
also marked that way.
}
\description{
Splitting lists of items
}
