% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecalcs.R
\name{tablecalcs}
\alias{tablecalcs}
\alias{tableNrow}
\alias{tableNcol}
\alias{tableDim}
\title{Calculations on tables}
\usage{
tableNrow(table)

tableNcol(table)

tableDim(table)
}
\arguments{
\item{table}{A known tabular-like environment object.}
}
\value{
\code{tableNrow()} returns the number of rows in the table.

\code{tableNcol()} returns the number of columns in the table.

\code{tableDim()} returns the number of rows and columns in the table.
}
\description{
Calculations on tables
}
\examples{
latex <- kableExtra::kbl(mtcars[1:2, 1:3], format = "latex")
parsed <- parseLatex(latex)
table <- parsed[[find_tabular(parsed)]]
table
tableNrow(table)
tableNcol(table)
tableDim(table)
}
