% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables}
\alias{tables}
\alias{is_Tabular}
\alias{find_tabular}
\alias{find_posOption}
\alias{posOption}
\alias{posOption<-}
\alias{find_widthOption}
\alias{widthOption}
\alias{widthOption<-}
\alias{find_columnOptions}
\alias{columnOptions}
\alias{columnOptions<-}
\alias{find_tableContent}
\alias{tableContent}
\alias{tableContent<-}
\title{Functions related to parsing LaTeX tables}
\usage{
is_Tabular(item)

find_tabular(items, start = 1)

find_posOption(table)

posOption(table)

posOption(table, asis = FALSE) <- value

find_widthOption(table)

widthOption(table)

widthOption(table, asis = FALSE) <- value

find_columnOptions(table)

columnOptions(table)

columnOptions(table, asis = FALSE) <- value

find_tableContent(table)

tableContent(table)

tableContent(table, asis = FALSE) <- value
}
\arguments{
\item{item}{An item from a LaTeX2 list object.}

\item{items}{A LaTeX2 list object.}

\item{start}{Where to start looking}

\item{table}{A known tabular-like environment object.}

\item{asis}{Whether to make small modifications in replacement functions}

\item{value}{A character string or LaTeX2 object}
}
\value{
`is_Tabular()` returns boolean indicating if this is
a tabular-like environment.

`find_tabular()` returns the index of the first
tabular-like environment, or `NA` if none is found

`find_posOption()` returns the indices of the
entries corresponding to the "pos" option, including the
brackets, within the contents list (i.e. `table[[2]]`).

`posOption()` returns a LaTeX2 object containing the
"pos" option.

`find_widthOption()` returns the index of the
block corresponding to the "width" option, if there is one.
Only some tabular-like environments have these.

`widthOption()` returns a LaTeX2 object containing the
"width" option, if the table has one.

`find_columnOptions()` returns the index of the
block corresponding to the column spec.

`columnOptions()` returns a LaTeX2 object containing the
"columb" options.

`find_tableContent()` returns the indices of the
entries corresponding to content of the table.

`tableContent()` returns a LaTeX2 object containing
all of the table content after the options.
}
\description{
Functions related to parsing LaTeX tables
}
\details{
Unless `asis == TRUE`, the value for `value` in `posOption(table) <- value`
can be specified with or without the enclosing brackets.

Unless `asis = TRUE`, `tableContent(table) <- value`
 will add newlines
at the start end end if not present, to make the result
more readable.
}
\examples{
latex <- kableExtra::kbl(mtcars[1:2, 1:2], format = "latex")
parsed <- parseLatex(latex)
is_Tabular(parsed[[2]])


find_tabular(parsed)
table <- parsed[[find_tabular(parsed)]]
table

find_posOption(table)
table[[2]][find_posOption(table)]

posOption(table)

posOption(table) <- "h"
posOption(table)
find_widthOption(table)

widthOption(table)

find_columnOptions(table)
columnOptions(table)

columnOptions(table) <- "lrr"
table
find_tableContent(table)

tableContent(table)

tableContent(table) <- "Mazda RX4 & 21 & 6\\\\\\\\"
table
tableContent(table, asis = TRUE) <- "Mazda RX4 & 21 & 6\\\\\\\\"
table

}
