% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests}
\alias{tests}
\alias{is_env}
\alias{is_macro}
\alias{is_block}
\alias{is_bracket}
\alias{is_whitespace}
\title{Test objects}
\usage{
is_env(item, envtypes = NULL)

is_macro(item, macros = NULL)

is_block(item)

is_bracket(item, bracket)

is_whitespace(item)
}
\arguments{
\item{item}{An object of class "LaTeX2item" to test.}

\item{envtypes}{Types of Latex environment to check for,
e.g. `"table"`.}

\item{macros}{Which macros to match, e.g. `"\\\\caption"`.}

\item{bracket}{Which bracket are we looking for?}
}
\value{
`is_env()` returns a boolean if the item matches.

`is_macro()` returns a boolean indicating the match.

`is_block()` returns a boolean indicating whether the `item` is a block wrapped in curly braces.

`is_bracket()` returns a boolean indicating that the `item` is a bracket of the
specified type.

`is_whitespace()` returns a boolean indicating if the
`item` is a space, tab or newline.
}
\description{
Test objects
}
\examples{
is_bracket(parseLatex("[]")[[1]], "[")
}
