% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableRow.R
\name{vector_to_latex2}
\alias{vector_to_latex2}
\title{Convert vector to table row}
\usage{
vector_to_latex2(x)
}
\arguments{
\item{x}{A list or vector to convert.}
}
\value{
A LaTeX2 object containing the entries
of \code{x} concatenated.
}
\description{
Convert vector to table row
}
\examples{
vector_to_latex2(1:3)
}
