% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableRow.R
\name{vector_to_row}
\alias{vector_to_row}
\title{Convert vector to table row}
\usage{
vector_to_row(cells, asis = FALSE, linebreak = TRUE)
}
\arguments{
\item{cells}{A list or vector of cell contents.}

\item{asis}{If \code{FALSE}, add blanks around cell contents.}

\item{linebreak}{If \code{TRUE}, add a line break marker.}
}
\value{
A LaTeX2 object which could be a row
in a tabular object.
}
\description{
Convert vector to table row
}
\examples{
vector_to_row(1:3)
}
