% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_LaTeX2.R
\name{as_LaTeX2}
\alias{as_LaTeX2}
\title{Coerce to LaTeX2}
\usage{
as_LaTeX2(x)
}
\arguments{
\item{x}{An object to convert to a LaTeX2 object.}
}
\value{
\code{as_LaTeX2()} converts \code{x} to a LaTeX2 object.
}
\description{
Coerce to LaTeX2
}
