% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLatex.R
\docType{data}
\name{defaultCatcodes}
\alias{defaultCatcodes}
\title{The default "catcodes" used by \link{parseLatex}.}
\format{
An object of class \code{data.frame} with 13 rows and 2 columns.
}
\usage{
defaultCatcodes
}
\description{
The default "catcodes" used by \link{parseLatex}.
}
\details{
\code{defaultCatcodes} is a dataframe containing the
default catcode definitions.
}
\examples{
# \makeatletter has no effect by default...
unclass(parseLatex("\\\\makeatletter\\\\internal@macro"))
# ... but the effect can be simulated
atletter <- rbind(defaultCatcodes,
                  data.frame(char="@", catcode=11))
unclass(parseLatex("\\\\makeatletter\\\\internal@macro",
                   catcodes = atletter))

}
\keyword{datasets}
