% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLatex.R
\name{deparseLatex}
\alias{deparseLatex}
\title{Convert latex object into character vector}
\usage{
deparseLatex(x, dropBraces = FALSE)
}
\arguments{
\item{x}{A latex object.}

\item{dropBraces}{Whether to drop unnecessary braces.}
}
\value{
\code{deparseLatex} returns character vector corresponding
to the parsed Latex.
#' @export
}
\description{
Convert latex object into character vector
}
