% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sequence.R
\name{is_equal}
\alias{is_equal}
\alias{find_sequence}
\title{Find a code sequence}
\usage{
is_equal(items1, items2)

find_sequence(items, sequence, all = FALSE, ignore_whitespace = TRUE)
}
\arguments{
\item{items1, }{items2 Two \link{LaTeX2} or \link{LaTeX2item} objects.}

\item{items, }{sequence \link{LaTeX2} objects or lists.}

\item{all}{Whether to return all matches, or just the first.}

\item{ignore_whitespace}{Whether to ignore whitespace in comparisons.}
}
\value{
\code{is_equal} returns a logical indicator of equality after removing source references.

Returns a path or list of paths where \code{sequence}
occurs within \code{items}.
}
\description{
Find a code sequence
}
\examples{
find_sequence(parseLatex("a & b & c"), "b & c")
}
