% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{find_whitespace}
\alias{find_whitespace}
\alias{find_env}
\alias{find_macro}
\alias{find_catcode}
\alias{find_char}
\title{Miscellaneous low-level finders}
\usage{
find_whitespace(items)

find_env(items, envtypes)

find_macro(items, macros)

find_catcode(items, codes)

find_char(items, char)
}
\arguments{
\item{items}{A list of latex items}

\item{envtypes}{Which types of environment to look for.}

\item{macros}{Which types of macros to look for.}

\item{codes}{Which codes to look for}

\item{char}{Which character to look for}
}
\value{
`find_whitespace()` returns the indices of
whitespace in `items`

`find_env()` returns the indices within `items`
of environments in `envtypes`.

`find_macro()` returns the index within `items`
of instances in `macros`.

`find_catcode()` returns the index within `items`
of specials matching `code`.

`find_char()` returns the index within `items`
of characters matching `char`.  Only characters
marked as SPECIAL by the parser will be found.
}
\description{
Miscellaneous low-level finders
}
