% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_contents}
\alias{get_contents}
\title{Convenience function to get contents from an item}
\usage{
get_contents(item)
}
\arguments{
\item{item}{An item from a Latex list (or a list with one item).}
}
\value{
`get_contents` returns the contents of the item as a Latex list, or as a
character string.
}
\description{
Convenience function to get contents from an item
}
\examples{
get_contents(parseLatex("{abc}"))

}
