% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{LaTeX2range}
\alias{LaTeX2range}
\alias{print.LaTeX2range}
\title{Ranges within LaTeX2 lists.}
\usage{
LaTeX2range(path, range)

\method{print}{LaTeX2range}(x, source = NULL, ...)
}
\arguments{
\item{path}{An integer vector to use as a path.}

\item{range}{A range of values within the path.}

\item{x}{Object to print.}

\item{source}{Optional parsed list from which to
extract the range.}

\item{...}{Ignored.}
}
\value{
\code{LaTeX2range()} returns a constructed \code{LaTeX2range} object.
}
\description{
Ranges within LaTeX2 lists.
}
\details{
LaTeX2range objects are lists with \code{path} and \code{range} entries.  \code{path} is a recursive index
into a LaTeX2 list, and \code{range} is a range of
entries in the result.

If \code{path} is \code{NULL}, the object refers to the entire
source object.  If \code{range} is \code{NULL}, it refers
to the whole \code{LaTeX2item} given by the \code{path}.
}
