% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{Names}
\alias{Names}
\alias{latexTag}
\alias{catcode}
\alias{envName}
\alias{macroName}
\title{Utility functions finding names and types of objects}
\usage{
latexTag(item)

catcode(item)

envName(item)

macroName(item)
}
\arguments{
\item{item}{A single latex item.}
}
\value{
`latexTag()` returns the LaTeX2 tag for the item or `NULL`.

`catcode()` returns the TeX catcode for the item, or `NULL`.

`envName()` returns the Latex environment name for an item, or `NULL`.

`macroName()` returns the Latex macro, or `NULL`.
}
\description{
Utility functions finding names and types of objects
}
