% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLatex.R
\name{parseLatex}
\alias{parseLatex}
\title{Parse LaTeX code}
\usage{
parseLatex(
  text,
  filename = deparse1(substitute(text)),
  verbose = FALSE,
  verbatim = c("verbatim", "verbatim*", "Sinput", "Soutput"),
  verb = "\\\\Sexpr"
)
}
\arguments{
\item{text}{A character vector containing LaTeX source code.}

\item{filename}{A filename to use in syntax error messages.}

\item{verbose}{If \code{TRUE}, print debug error messages.}

\item{verbatim}{A character vector containing the names of \LaTeX environments holding verbatim text.}

\item{verb}{A character vector containing LaTeX macros that should be assumed to hold verbatim text.}
}
\value{
`NULL`, so far.
}
\description{
The \code{parseLatex} function parses LaTeX source, producing a structured object.
}
\examples{
parsed <- parseLatex(r"(fran\c{c}ais)")
parsed
getParseData(parsed)
}
