% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLatex.R
\docType{package}
\name{parseLatex}
\alias{parseLatex-package}
\alias{parseLatex}
\alias{print.LaTeX2}
\title{Parse LaTeX code}
\usage{
parseLatex(
  text,
  filename = deparse1(substitute(text)),
  verbose = FALSE,
  verbatim = c("verbatim", "verbatim*", "Sinput", "Soutput"),
  verb = "\\\\Sexpr"
)

\method{print}{LaTeX2}(x, ...)
}
\arguments{
\item{text}{A character vector containing LaTeX source code.}

\item{filename}{A filename to use in syntax error messages.}

\item{verbose}{If \code{TRUE}, print debug error messages.}

\item{verbatim}{A character vector containing the names of \LaTeX environments holding verbatim text.}

\item{verb}{A character vector containing LaTeX macros that should be assumed to hold verbatim text.}
}
\value{
Parsed Latex in list with class `"LaTeX2"`.
}
\description{
The \code{parseLatex} function parses LaTeX source, producing a structured object.

This package exports an enhanced version of the tools::parseLatex function, which handles more LaTeX syntax.
}
\examples{
parsed <- parseLatex(r"(fran\c{c}ais)")
parsed
getParseData(parsed)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmurdoch/parseLatex}
  \item Report bugs at \url{https://github.com/dmurdoch/parseLatex/issues}
}

}
\author{
\strong{Maintainer}: Duncan Murdoch \email{murdoch.duncan@gmail.com}

}
