% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{path_to_index}
\alias{path_to_index}
\alias{index_to_path}
\title{Convert between paths and indices}
\usage{
path_to_index(path, items)

index_to_path(index, items)
}
\arguments{
\item{path}{A vector of integers, assumed to be
a path through "ITEMLIST" entries in a \link{LaTeX2}
or \link{LaTeX2item} object.}

\item{items}{The referenced object.}

\item{index}{A scalar integer which would be
the index to an item if \code{items} was flattened.}
}
\value{
\code{path_to_index} returns a scalar value
corresponding the the index if \code{items} was flattened.

\code{index_to_path} returns a vector of integers
which would index the specified item.
}
\description{
Convert between paths and indices
}
\seealso{
\code{\link[=flatten_itemlists]{flatten_itemlists()}}
}
