% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showErrors.R
\name{showErrors}
\alias{showErrors}
\title{Show errors in parsed Latex object}
\usage{
showErrors(x)
}
\arguments{
\item{x}{A \link{LaTeX2} object.}
}
\value{
A list of paths to errors, invisibly.
}
\description{
Show errors in parsed Latex object
}
\examples{
parsed <- parseLatex("\\\\begin{foo} abc \\\\end{bar} def",
                     recover = TRUE)
showErrors(parsed)
}
