% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOption.R
\name{find_posOption}
\alias{find_posOption}
\alias{posOption}
\alias{posOption<-}
\alias{find_widthOption}
\alias{widthOption}
\alias{widthOption<-}
\alias{find_columnOptions}
\alias{columnOptions}
\alias{columnOptions<-}
\title{Functions related to table options.}
\usage{
find_posOption(table)

posOption(table)

posOption(table, asis = FALSE) <- value

find_widthOption(table)

widthOption(table)

widthOption(table, asis = FALSE) <- value

find_columnOptions(table)

columnOptions(table)

columnOptions(table, asis = FALSE) <- value
}
\arguments{
\item{table}{A known tabular-like environment object.}

\item{asis}{Whether to make small modifications in replacement functions}

\item{value}{A character string or LaTeX2 object}
}
\value{
`find_posOption()` returns the indices of the
entries corresponding to the "pos" option, including the
brackets, within the contents list (i.e. `table[[2]]`).

`posOption()` returns a LaTeX2 object containing the
"pos" option.

`find_widthOption()` returns the index of the
block corresponding to the "width" option, if there is one.
Only some tabular-like environments have these.

`widthOption()` returns a LaTeX2 object containing the
"width" option, if the table has one.

`find_columnOptions()` returns the index of the
block corresponding to the column spec.

`columnOptions()` returns a LaTeX2 object containing the
"column" options.
}
\description{
Functions related to table options.
}
\details{
Unless `asis == TRUE`, the value for `value` in `posOption(table) <- value`
can be specified with or without the enclosing brackets.
}
\examples{
latex <- kableExtra::kbl(mtcars[1:2, 1:2], format = "latex")
parsed <- parseLatex(latex)
table <- parsed[[find_tabular(parsed)]]
table
find_posOption(table)

posOption(table)

posOption(table) <- "h"
posOption(table)
find_widthOption(table)

widthOption(table)

find_columnOptions(table)
columnOptions(table)

columnOptions(table) <- "lrr"
table
}
