% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableRow.R
\name{find_tableRow}
\alias{find_tableRow}
\alias{tableRow}
\alias{tableRow<-}
\title{Functions to work with rows in tables}
\usage{
find_tableRow(table, row)

tableRow(table, row)

tableRow(table, row, asis = FALSE) <- value
}
\arguments{
\item{table}{A tabular-like environment to work with.}

\item{row}{row in the table (1 is top row), including
rows of labels}

\item{asis}{Should a linebreak and newline be added after the
value?}

\item{value}{The content to be inserted into the cell.  This
can be a LaTeX2 object, or a character string that will be
converted to one.}
}
\value{
`find_tableRow()` returns the indices of the
entries corresponding to the content of row i of the table.

`tableRow()` returns a LaTeX2 object containing
all of the table content in the row
}
\description{
Functions to work with rows in tables
}
\details{
Unless `asis = TRUE`, `tableContent(table) <- value`
 will add "\\" and a newline
at the end if not present.

If the `row` value is higher than the number of rows
in the table, blank rows will be added to fill the
space between.
}
\examples{
latex <- kableExtra::kbl(mtcars[1:2, 1:2], format = "latex")
parsed <- parseLatex(latex)
table <- parsed[[find_tabular(parsed)]]
find_tableRow(table, 1)

tableRow(table, 1)

tableRow(table, 5) <- "a & b & c"
table

}
