% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{drop_items}
\alias{select_items}
\alias{drop_whitespace}
\alias{include_whitespace}
\alias{split_list}
\alias{split_latex}
\title{Miscellaneous utilities}
\usage{
drop_items(items, which)

select_items(items, which)

drop_whitespace(items)

include_whitespace(items, which)

split_list(items, splits)

split_latex(items, splits)
}
\arguments{
\item{items}{A LaTeX2 object or list of items.}

\item{which}{Which items to operate on.}

\item{splits}{Which items divide the parts?}
}
\value{
`drop_items()` returns the list of items with specific items removed.

`select_items()` returns the list of subsetted items

`drop_whitespace()` returns the list of items with whitespace (blanks, tabs, newlines) removed.

`include_whitespace()` returns `which` with
following whitespace (blanks, tabs, newlines) included.

`split_list()` returns a list of pieces
separated at the splits

`split_latex()` returns a list of pieces
separated at the splits
}
\description{
Miscellaneous utilities
}
