% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Utilities}
\alias{Utilities}
\alias{latexTag}
\alias{catcode}
\alias{envName}
\alias{macroName}
\alias{is_env}
\alias{is_macro}
\alias{is_block}
\alias{as_LaTeX2}
\alias{is_bracket}
\alias{is_whitespace}
\alias{drop_items}
\alias{select_items}
\alias{drop_whitespace}
\alias{find_env}
\alias{find_macro}
\alias{find_catcode}
\alias{split_items}
\title{Utility functions for working with parsed Latex.}
\usage{
latexTag(item)

catcode(item)

envName(item)

macroName(item)

is_env(item, envtypes = NULL)

is_macro(item, macros = NULL)

is_block(item)

as_LaTeX2(items)

is_bracket(item, bracket)

is_whitespace(item)

drop_items(items, which)

select_items(items, which)

drop_whitespace(items)

find_env(items, envtypes)

find_macro(items, macros)

find_catcode(items, codes)

split_items(items, splits)
}
\arguments{
\item{item}{A single latex item.}

\item{envtypes}{Types of Latex environment to check for,
e.g. `"table"`.}

\item{macros}{Which macros to match, e.g. `"\\\\caption"`.}

\item{bracket}{Which bracket are we looking for?}

\item{codes}{Which codes to look for}

\item{splits}{Which items divide the parts?}
}
\value{
`latexTag()` returns the LaTeX2 tag for the item or `NULL`.

`catcode()` returns the TeX catcode for the item, or `NULL`.

`envName()` returns the Latex environment name for an item, or `NULL`.

`macroName()` returns the Latex macro, or `NULL`.

`is_env()` returns a boolean if the item matches.

`is_macro()` returns a boolean indicating the match.

`is_block()` returns a boolean indicating whether the `item` is a block wrapped in curly braces.

`as_LaTeX2()` marks the `items` as parsed Latex so that
they print nicely.

`is_bracket()` returns a boolean indicating that the `item` is a bracket of the
specified type.

`is_whitespace()` returns a boolean indicating if the
`item` is a space, tab or newline.

`drop_items()` returns the list of items with specific items removed.

`select_items()` returns the list of subsetted items

`drop_whitespace()` returns the list of items with whitespace (blanks, tabs, newlines) removed.

`find_env()` returns the indices within `items`
of environments in `envtypes`.

`find_macro()` returns the index within `items`
of instances in `macros`.

`find_catcode()` returns the index within `items`
of specials matching `code`.

`split_items()` returns a list of pieces
separated at the splits
}
\description{
Utility functions for working with parsed Latex.
}
\examples{
is_bracket(parseLatex("[]")[[1]], "[")
}
