% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{latexTag}
\alias{latexTag}
\alias{catcode}
\alias{envname}
\alias{macroname}
\alias{is_bracket}
\alias{dropWhitespace}
\title{Utility functions for working with parsed Latex.}
\usage{
latexTag(item)

catcode(item)

envname(item)

macroname(item)

is_bracket(item, bracket)

dropWhitespace(items)
}
\arguments{
\item{item}{A single latex item.}

\item{bracket}{Which bracket are we looking for?}
}
\value{
`latexTag()` returns the LaTeX2 tag for the item or `NULL`.

`catcode()` returns the TeX catcode for the item, or `NULL`.

`envname()` returns the Latex environment name for an item, or `NULL`.

`macroname()` returns the Latex macro, or `NULL`.

`is_bracket()` returns a boolean indicating that the `item` is a bracket of the
specified type.

`dropWhitespace()` returns the list of items with whitespace (blanks, tabs, newlines) removed.
}
\description{
Utility functions for working with parsed Latex.
}
\examples{
is_bracket(parseLatex("[]")[[1]], "[")
}
